Fix the same int accumulator narrowing bug across 6 tps/fps/auc sites in AUC kernels
test_fairness_roc_auc_uses_positive_class_indicator failed under NUMBA_DISABLE_JIT=1 with a
nonsensical AUC (-1.4375). Root cause: the same class of bug just fixed in
_batch_per_class_ice_kernel (_classification_report.py) -- tps/fps/auc/last_counted_* accumulators
seeded from the bare Python literal 0, accumulating narrow-dtype indicator elements
(fast_numba_auc_nonw and its siblings assume 0/1 y_true). numpy's runtime value-based casting
narrows the running total to the array's dtype once it first becomes a numpy scalar, silently
wrapping under normal numba JIT (type inference widens differently there) but raising/corrupting
under NUMBA_DISABLE_JIT=1's pure-Python/numpy execution.

A repo-wide grep for the same "tps = 0" / "tps, fps = 0, 0" pattern found 6 occurrences across
_core_auc_brier.py (fast_numba_auc_nonw and 5 siblings) plus 2 more in
_bootstrap_fused_binary_bundle.py and 1 in a benchmark script -- fixed all 9 by seeding every
accumulator from np.int64(0). 3 return sites needed an explicit float(...) wrap (mypy: np.int64
division loses static float typing). Verified: 170 tests across tests/evaluation, tests/metrics,
and the fairness test all pass unchanged under normal JIT; the 3 previously-failing fairness cases
now also pass under NUMBA_DISABLE_JIT=1.

+++ src/mlframe/evaluation/_bootstrap_fused_binary_bundle.py
@@ -86,11 +86,11 @@ def _bootstrap_batch_auc_brier_ll_ece(
             rk = base_rank[idx[k]]
             counts[rk] += 1
             ones[rk] += y_by_rank[rk]
-        last_fps = 0
-        last_tps = 0
-        tps = 0
-        fps = 0
-        auc = 0
+        last_fps = np.int64(0)
+        last_tps = np.int64(0)
+        tps = np.int64(0)
+        fps = np.int64(0)
+        auc = np.int64(0)
         for rk in range(n - 1, -1, -1):
             c = counts[rk]
             if c == 0:
@@ -148,11 +148,11 @@ def _bootstrap_batch_auc_brier_ll_ece_grouped(
             g = group_of_base[bi]
             counts[g] += 1
             ones[g] += y_base[bi]
-        last_fps = 0
-        last_tps = 0
-        tps = 0
-        fps = 0
-        auc = 0
+        last_fps = np.int64(0)
+        last_tps = np.int64(0)
+        tps = np.int64(0)
+        fps = np.int64(0)
+        auc = np.int64(0)
         for g in range(ngroups - 1, -1, -1):
             c = counts[g]
             if c == 0:
